% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dict.R
\name{create_dict}
\alias{create_dict}
\title{Crée un dictionnaire à la manière d'un dict Python}
\usage{
create_dict(keys, values)
}
\arguments{
\item{keys}{Vecteur de chaînes de caractères représentant les clés du dictionnaire.}

\item{values}{Vecteur de valeurs associées aux clés du dictionnaire.}
}
\value{
Un tibble représentant le dictionnaire avec les clés et les valeurs.
}
\description{
Cette fonction crée un dictionnaire en mettant en relation les keys et values.
}
