% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_price_range.R
\name{create_price_range}
\alias{create_price_range}
\title{Crée une variable de plage de prix à partir du prix de vente au détail.}
\usage{
create_price_range(df_prices)
}
\arguments{
\item{df_prices}{Le dataframe contenant les prix de vente au détail.}
}
\value{
Le dataframe avec une nouvelle variable 'price_range' indiquant la plage de prix pour chaque prix de vente au détail.
}
\description{
Cette fonction prend un dataframe contenant les prix de vente au détail et crée une nouvelle variable indiquant la plage de prix à laquelle chaque prix appartient.
}
\examples{
df <- tibble::tibble(retail_price = c(3.99, 15.50, 30.25, 99.99))
create_price_range(df)

}
