% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values.R
\name{extract_values}
\alias{extract_values}
\title{Fait la requête et extrait les informations pour obtenir un dictionnaire}
\usage{
extract_values(link, max_retry, rowid)
}
\arguments{
\item{link}{Lien vers le site BrickEconomy à requêter.}

\item{max_retry}{Nombre maximum de tentatives de requête en cas d'échec.}

\item{rowid}{Identifiant de ligne pour un suivi clair des liens scrapés.}
}
\value{
Un tibble représentant le dictionnaire avec les noms de variable et les valeurs.
}
\description{
Cette fonction fait la requête sur le site BrickEconomy en utilisant le lien
fourni, puis extrait les informations nécessaires pour obtenir un dictionnaire
avec les noms de variable et les valeurs associées. Elle dépend des autres fonctions
telles que \code{make_request}, \code{get_keys}, \code{get_values}, et \code{create_dict}.
}
