% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_reviews.R
\name{get_all_reviews}
\alias{get_all_reviews}
\title{Récupère tous les avis pour tous les ensembles présents dans le dataframe}
\usage{
get_all_reviews(df_sets, type)
}
\arguments{
\item{df_sets}{Dataframe contenant des informations sur les ensembles.}

\item{type}{Type d'avis à récupérer ('ratings', 'text', 'all').}
}
\value{
Un dataframe élargi avec les avis spécifiés pour chaque ensemble.
}
\description{
Cette fonction prend un dataframe contenant des informations sur plusieurs
ensembles et permet de récupérer tous les avis, soit les notes, les avis textuels,
ou tous les types d'avis, pour chaque ensemble. Le type d'avis est spécifié
par le paramètre 'type' (choisir parmi 'ratings', 'text', 'all'). Assurez-vous
d'avoir préalablement enregistré la clé API, le nom d'utilisateur et le mot
de passe dans les options.
}
\seealso{
\code{\link{get_only_ratings}}, \code{\link{get_only_text}}, \code{\link{get_reviews}}
}
