% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keys.R
\name{get_keys}
\alias{get_keys}
\title{Récupère les noms des données scrapées}
\usage{
get_keys(response)
}
\arguments{
\item{response}{Objet de réponse de la requête HTTP.}
}
\value{
Vecteur de chaînes de caractères contenant les noms des données.
}
\description{
Cette fonction extrait les noms des données scrapées à partir de la réponse
de la requête HTTP.
}
