% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_only_ratings.R
\name{get_only_ratings}
\alias{get_only_ratings}
\title{Récupère uniquement les notes pour un ensemble}
\usage{
get_only_ratings(setID, reviewCount)
}
\arguments{
\item{setID}{Identifiant de l'ensemble pour lequel récupérer les notes.}

\item{reviewCount}{Nombre total d'avis disponibles pour cet ensemble.}
}
\value{
Un tableau contenant les moyennes des notes spécifiques.
}
\description{
Cette fonction utilise l'API de Brickset pour récupérer les notes spécifiques
d'un ensemble identifié par son setID. Les notes incluent la moyenne générale,
la moyenne des pièces, l'expérience de construction, la jouabilité et la valeur
pour l'argent. Assurez-vous d'avoir préalablement enregistré la clé API, le
nom d'utilisateur et le mot de passe dans les options.
}
