% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_only_text.R
\name{get_only_text}
\alias{get_only_text}
\title{Récupère uniquement les avis textuels pour un ensemble}
\usage{
get_only_text(setID, reviewCount)
}
\arguments{
\item{setID}{Identifiant de l'ensemble pour lequel récupérer les avis textuels.}

\item{reviewCount}{Nombre total d'avis disponibles pour cet ensemble.}
}
\value{
Un tableau contenant les avis textuels.
}
\description{
Cette fonction utilise l'API de Brickset pour récupérer les avis textuels
d'un ensemble identifié par son setID. Les informations incluses sont
l'auteur, la date de publication, le titre, le texte de l'avis et le format
HTML de l'avis. Assurez-vous d'avoir préalablement enregistré la clé API, le
nom d'utilisateur et le mot de passe dans les options.
}
