% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_instructions.R
\name{get_set_instructions}
\alias{get_set_instructions}
\title{Obtient les instructions d'assemblage pour un set donné.}
\usage{
get_set_instructions(ID)
}
\arguments{
\item{ID}{L'identifiant de l'ensemble pour lequel vous souhaitez obtenir les instructions d'assemblage.}
}
\value{
Une liste d'URLs vers les instructions d'assemblage au format HTML.
}
\description{
Cette fonction récupère les instructions d'assemblage pour un ensemble donné à partir de l'API Brickset.
}
\examples{
get_set_instructions("ID_du_set")

}
\references{
Pour plus d'informations sur l'API Brickset : \url{https://brickset.com/tools/webservices/v3}.
}
