% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Récupère les valeurs scrapées}
\usage{
get_values(response)
}
\arguments{
\item{response}{Objet de réponse de la requête HTTP.}
}
\value{
Vecteur de chaînes de caractères contenant les valeurs scrapées.
}
\description{
Cette fonction extrait les valeurs scrapées à partir de la réponse de la requête HTTP.
}
