% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_all_feathers.R
\name{import_all_feathers}
\alias{import_all_feathers}
\title{Importe tous les fichiers Feather disponibles dans le dossier par défaut}
\usage{
import_all_feathers(chemin_data = "data")
}
\arguments{
\item{chemin_data}{Chemin du dossier contenant les fichiers Feather. Par défaut : "data".}
}
\value{
Un dataframe regroupant toutes les données des fichiers Feather.
}
\description{
Cette fonction importe tous les fichiers Feather (.feather) disponibles dans le
dossier spécifié par défaut (par défaut, le dossier "data") et les regroupe tous
dans un seul dataframe. Assurez-vous que les fichiers Feather sont présents dans
le dossier spécifié.
}
