% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_all_parquets.R
\name{import_all_parquets}
\alias{import_all_parquets}
\title{Importe tous les fichiers .parquet en un seul dataframe}
\usage{
import_all_parquets(chemin_dossier = "data")
}
\arguments{
\item{chemin_dossier}{Chemin du dossier contenant les fichiers .parquet. Par défaut : "data".}
}
\value{
Un dataframe contenant toutes les données combinées.
}
\description{
Cette fonction combine tous les fichiers .parquet présents dans le dossier spécifié
(par défaut, le dossier "data") en un seul dataframe. Les fichiers
.parquet sont lus avec la fonction read_parquet du package arrow. Le dataframe
résultant est filtré pour inclure uniquement les catégories "Normal", "Extended" et
"Collection". Une colonne supplémentaire set_number est créée en combinant les
colonnes number et numberVariant, et une colonne brickeconomy_link est générée en
utilisant le package glue pour construire les liens correspondants sur le site BrickEconomy.
}
