% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_assembler.R
\name{lego_assembler}
\alias{lego_assembler}
\title{Joint le dataframe des sets avec le dataframe des prix scrapé des sets}
\usage{
lego_assembler(df_sets, df_scrap)
}
\arguments{
\item{df_sets}{Dataframe contenant des informations sur les ensembles.}

\item{df_scrap}{Dataframe contenant des informations sur les prix des ensembles.}
}
\value{
Un dataframe résultant de la jointure avec les colonnes supplémentaires.
}
\description{
Cette fonction effectue une jointure interne (inner join) entre le dataframe des sets
(`df_sets`) et le dataframe des prix des sets (`df_scrap`). La jointure est réalisée
en utilisant la colonne "set_number" du dataframe des sets et la colonne "brickeconomy_link"
du dataframe des prix des sets. Les enregistrements sans prix de détail sont supprimés.
La fonction ajoute également des colonnes supplémentaires, telles que "diff_new_retail"
(rapport entre la valeur en neuf et le prix de détail), "diff_used_retail" (rapport entre
la valeur en occasion et le prix de détail), et "ppp" (prix par pièce).
}
