% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{make_request}
\alias{make_request}
\title{Effectue une requête sur le site BrickEconomy}
\usage{
make_request(link, max_retry, rowid)
}
\arguments{
\item{link}{Lien vers le site BrickEconomy à requêter.}

\item{max_retry}{Nombre maximum de tentatives de requête en cas d'échec.}

\item{rowid}{Identifiant de ligne pour un suivi clair des liens scrapés.}
}
\value{
Objet de réponse de la requête HTTP.
}
\description{
Cette fonction effectue une requête sur le site BrickEconomy en utilisant le lien
fourni. Elle prend en charge des options telles que le nombre maximum de tentatives
de requête en cas d'échec (\code{max_retry}) et l'identifiant de ligne (\code{rowid}) pour
un suivi clair des liens scrapés. Assurez-vous d'avoir préalablement enregistré
la clé API, le nom d'utilisateur et le mot de passe dans les options.
}
