% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_cleaner.R
\name{price_cleaner}
\alias{price_cleaner}
\title{Nettoie une colonne de chaînes de caractères représentant des prix}
\usage{
price_cleaner(column)
}
\arguments{
\item{column}{Colonne de chaînes de caractères représentant des prix.}
}
\value{
Colonne nettoyée avec les prix transformés en double.
}
\description{
Cette fonction enlève le symbole €, remplace les virgules par des points,
supprime les espaces inutiles, remplace les valeurs "Unknown" par NA, et
transforme la colonne en type double.
}
