% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_table.R
\name{theme_table}
\alias{theme_table}
\title{Tableau interactif avec filtre par thème}
\usage{
theme_table(df_prices, theme)
}
\arguments{
\item{df_prices}{Le dataframe contenant les informations sur les prix des sets LEGO.}

\item{theme}{Le thème à filtrer dans le tableau.}
}
\description{
Cette fonction crée un tableau interactif avec un filtre par thème à partir du dataframe
fourni. Elle utilise le package gt pour générer une table HTML interactive avec des
fonctionnalités telles que le filtrage, le tri et la mise en forme des données. La table
affiche les informations spécifiques sur les sets LEGO en fonction du thème choisi.
}
